pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-pipeline-v2"
version = "0.1.0"
description = "Local-first Markdown to Chroma ingestion pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.9",
  "rich>=13.0",
  "PyYAML>=6.0",
  "chromadb>=0.4",
  "sentence-transformers>=2.2",
]

[project.scripts]
pipeline = "pipeline.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
